+++ src/pushMetrics.js
@@ -24,6 +24,8 @@ const probeLabelNames = [
   'socket_src_address',
   'socket_dst_family',
   'socket_dst_address',
+  'dns_resolvers',
+  'probe_name',
   'res_status',
   'err_code',
 ];