+++ test/pushMetricsTest.js
@@ -67,7 +67,9 @@ describe('pushMetrics', () => {
         duration: 140,
         probe_status: 0,
         probe_id: 'https://example.com/foo',
-        probe_env: 'producation'
+        probe_env: 'producation',
+        dns_resolvers: ['8.8.8.8', '9.9.9.9'],
+        probe_label: 'custom-dns'
       },
       {
         protocol: 'dns',
@@ -84,13 +86,13 @@ describe('pushMetrics', () => {
     assert.equal(`# HELP lastmile_probe_duration_milliseconds duration of the request from lastmile
 # TYPE lastmile_probe_duration_milliseconds gauge
 lastmile_probe_duration_milliseconds{probe_status="1",instance="hostname_i1",instance_address="192.0.0.3",probe_id="probeid",probe_failed_expectations="F1,F2",probe_env="development",err_code="ERR_FOO"} 1337
-lastmile_probe_duration_milliseconds{probe_status="0",instance="hostname_i1",instance_address="192.0.0.3",protocol="https",req_url="https://example.com/foo",req_host="example.com",req_port="443",req_path="/foo",req_method="GET",probe_id="https://example.com/foo",probe_env="producation",socket_tls_protocol="TLSv3",socket_src_address="172.0.0.3",socket_dst_family="IPv4",socket_dst_address="52.0.0.3",res_status="200"} 140
+lastmile_probe_duration_milliseconds{probe_status="0",instance="hostname_i1",instance_address="192.0.0.3",protocol="https",req_url="https://example.com/foo",req_host="example.com",req_port="443",req_path="/foo",req_method="GET",probe_id="https://example.com/foo",probe_env="producation",socket_tls_protocol="TLSv3",socket_src_address="172.0.0.3",socket_dst_family="IPv4",socket_dst_address="52.0.0.3",dns_resolvers="8.8.8.8,9.9.9.9",probe_name="custom-dns",res_status="200"} 140
 lastmile_probe_duration_milliseconds{probe_status="0",instance="hostname_i1",instance_address="192.0.0.3",protocol="dns",res_addresses="127.0.0.1,127.0.0.2",probe_env="producation"} 140
 
 # HELP lastmile_probe_status probe status (0=ok, 1=error)
 # TYPE lastmile_probe_status gauge
 lastmile_probe_status{instance="hostname_i1",instance_address="192.0.0.3",probe_id="probeid",probe_failed_expectations="F1,F2",probe_env="development",err_code="ERR_FOO"} 1
-lastmile_probe_status{instance="hostname_i1",instance_address="192.0.0.3",protocol="https",req_url="https://example.com/foo",req_host="example.com",req_port="443",req_path="/foo",req_method="GET",probe_id="https://example.com/foo",probe_env="producation",socket_tls_protocol="TLSv3",socket_src_address="172.0.0.3",socket_dst_family="IPv4",socket_dst_address="52.0.0.3",res_status="200"} 0
+lastmile_probe_status{instance="hostname_i1",instance_address="192.0.0.3",protocol="https",req_url="https://example.com/foo",req_host="example.com",req_port="443",req_path="/foo",req_method="GET",probe_id="https://example.com/foo",probe_env="producation",socket_tls_protocol="TLSv3",socket_src_address="172.0.0.3",socket_dst_family="IPv4",socket_dst_address="52.0.0.3",dns_resolvers="8.8.8.8,9.9.9.9",probe_name="custom-dns",res_status="200"} 0
 lastmile_probe_status{instance="hostname_i1",instance_address="192.0.0.3",protocol="dns",res_addresses="127.0.0.1,127.0.0.2",probe_env="producation"} 0
 
 # HELP lastmile_last_seen timestamp